+++ src/test/java/qna/domain/AnswerRepositoryTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import qna.config.QnaDataJpaTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static qna.domain.AnswerTest.A1;
+import static qna.domain.FixtureAnswer.A1;
 
 @QnaDataJpaTest
 class AnswerRepositoryTest {
