+++ src/test/java/qna/domain/FixtureUser.java
@@ -1,6 +1,9 @@
 package qna.domain;
 
-public class UserTest {
+public final class FixtureUser {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
     public static final User SANJIGI = new User(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
+
+    private FixtureUser() {
+    }
 }
